+++ src/main/java/org/opensearch/neuralsearch/processor/chunker/Chunker.java
@@ -54,4 +54,6 @@ public interface Chunker {
     static boolean checkRunTimeMaxChunkLimit(int chunkResultSize, int runtimeMaxChunkLimit, int chunkStringCount) {
         return runtimeMaxChunkLimit != DISABLED_MAX_CHUNK_LIMIT && chunkResultSize + chunkStringCount >= runtimeMaxChunkLimit;
     }
+
+    String getAlgorithmName();
 }